+++ packages/stable/src/types.ts
@@ -34,7 +34,7 @@ export {
   AutoPagingToArray,
 } from '@cognite/sdk-core';
 
-export * from './api/spatial/types';
+export * from './api/geospatial/types';
 
 export interface Acl<ActionsType, ScopeType> {
   actions: ActionsType[];