+++ packages/beta/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.10](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.0.9...@cognite/sdk-beta@5.0.10) (2022-01-10)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.0.9](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.0.8...@cognite/sdk-beta@5.0.9) (2022-01-10)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
 ## [5.0.8](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.0.7...@cognite/sdk-beta@5.0.8) (2021-12-21)
 
 **Note:** Version bump only for package @cognite/sdk-beta